// src/app/models/sjcs/JuzgadoItem.ts
export class JuzgadoItem {
  idJuzgado: string;
  nombre: string;
  domicilio: string;
  idInstitucion: string;
  fechaModificacion: string;
  codigoPostal: string;
  idPoblacion: string;
  idProvincia: string;
  historico: boolean;
  fechabaja: string;
  telefono1: string;
  telefono2: string;
  fax: string;
  codigoExt: string;
  codigoProcurador: string;
  visible: string;
  movil: string;
  esDecano: string;
  codigoExt2: string;
  email: string;
  isCodigoEjis: string;
  visibleMovil: string;
  fechaCodigoEjis: any;
  nombrePoblacion: string;
  nombreProvincia: string;
  constructor() { }
}